Fix data corupption after failing to upload a truncated file
If the user's IAM role does not have permission to write to the S3
Bucket, the original data will be lost after the file is overwritten.

This is because overwrite writing will be divided into two steps:
truncation and writing. First, truncating the file will succeed, but
uploading will fail, so subsequent writing will not happen. But at this
time, the timestamp of the local file is newer than that on the server,
so S3FS will think that the local data is correct when it is read next
time, but the old data on the server is actually correct.

This patch will clear the timestamp of the local cache to invalidate the
cache after the upload fails, and force S3FS to re-download the data at
the next read to fix this problem.

Signed-off-by: Qinqi Qu <quqinqi@linux.alibaba.com>